feat: refactor delete method

+++ src/main/java/com/apiproducts/products/services/ProductService.java
@@ -1,12 +1,17 @@
 package com.apiproducts.products.services;
 
+import com.apiproducts.products.controllers.ProductController;
 import com.apiproducts.products.models.Product;
 import com.apiproducts.products.repositories.ProductRepository;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
+
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
 @Service
 public class ProductService {
@@ -31,7 +36,13 @@ public class ProductService {
     }
 
     @Transactional
-    public void delete(Product product) {
-        productRepository.delete(product);
+    public ResponseEntity delete(String id) {
+        Optional<Product> optionalProduct = productRepository.findById(id);
+        if(optionalProduct.isPresent()){
+            Product product = optionalProduct.get();
+            productRepository.delete(product);
+            return ResponseEntity.ok("Produto deletado com sucesso!");
+        }
+        return ResponseEntity.notFound().build();
     }
 }
